refactor: Extract Client logic into Smith trait
